Add a text-mode renderer.
I don't expect this to be part of the "real game", but I do expect it to
be useful for e.g. test diagnostic output, and for now this seemed
easier to get started with than integrating OpenGL/WebGL (which is still
planned).

There does appear to be something wrong with the raycaster when it deals
with nearly-axis-aligned lines, as the display flashes blank then.