Enable overflow checks in release mode.
Not having overflow checks seems like a bad idea on general principle:
large numbers are more likely to come up in production than testing.
I think I'm programming with care for numeric ranges, but I'm highly
likely to be wrong...

That said, this is more of a philosophical point than practical right
now, since this is nowhere near a release.