Add wasm-bindgen-test test of some of gui.js and the bindings.
This is mostly an experimental setup of testing; the real value will be
in actually automated tests of the JS/WASM glue code (i.e.
`gameapp.rs`). I wrote this code a long time ago but didn't get it to
work back when I was using npm & webpack, not solely wasm-pack.