Move ray projection logic into ProjectionHelper.
ProjectionHelper now also stores the view matrix; this is slightly
untidy from a mutability perspective but allows caching of all the
needed derived matrices.

Motivation: We're going to want to reuse project_ndc_into_world
for user interaction with blocks, so it should be moved out of the
raytracing-specific code.