Put luminance-front back, and move some WebGL setup.
I previously misunderstood what I had to do to make the luminance-using
code more general. luminance-front is actually helpful and we only need
to keep the actual WebGL-specific things in the WebGL land.

This commit moves all of the external WebGL-specific stuff out of
glrender.js; unfortunately it adds more references to CanvasHelper,
but those can be abstracted later.