Make Face lists into constants; reorder variants.
Changing Face::all_six() into Face::ALL_SIX will hopefully make it a
little easier for the compiler to make things efficient if it sees a
good opportunity.

Reordering the variants is just because it feels correct for the most
"undefined"-like variant (WITHIN) to be the first one in the list.