Modify TextureAllocator trait to support texture atlases.
This does not implement allocating tiles in an atlas; it only adds
the necessary interface to scale and offset texture coordinates.

It turns out that on my desktop GPU, the maximum array size is 2048,
which is not going to last long when my current hello-world tinkering
needs 966 tiles. Moving to an atlas will let us have (4096 / 18)² = 50k
tiles in just one layer (16×16 tiles with border).