Add a Result return value to Space::set().
The motivation is to enable set() to not panic when there is a problem
with a Block's interior URef, but arranging to make that happen turned
out tricky so I'm doing smaller commits.

The other cases of the return value type are intended to be eventually
useful for UI feedback ("why did placing/destroying this block fail?").