Remove Block::space method.
I'm expecting to change the structure of Block such that it will become
more important to use Block::evaluate for reading/rendering, and there
will be more possible cases than "this block directly refers to a
Space". Thus, space() is a poor abstraction.

So are some other Block methods, so I've added notes on why they're
problematic.