Implement continuous collision detection and sliding.
There is a bug in this version of the algorithm, such that pushing
against an inside corner will eventually, but not immediately, allow
passing through one of the blocks. I am letting it be for the moment
expecting that future experimentation will reveal the problem.

Also, this version does not reset velocities to zero on collision.
(Fixing that might end up hiding the above bug.)