Implement volume collision.
This has been put off for a while because there have been serious bugs,
and there is still at least one: sliding against a wall by moving almost
perpendicular to it allows passing through a corner in the off-axis
direction. But I think that we can squash that later, and this collision
implementation is solid enough to start adding gravity without it being
an inconvenience.