Make VoxelDisplayAdapter use a 3D origin instead of a Z coordinate.
This simplifies existing code that takes an origin point, and I expect
it to be useful in future code using drawing. (Eventually we'll want
general rotations as well as translations, but I expect to implement
that for all `Space` manipulation rather than 2D drawing in specific.)

I also removed some `as` expressions that are no longer necessary since
`GridCoordinate` is always `i32` instead of `isize` now.