Make `Space::extract` and `Space::distinct_blocks_unfiltered_iter` return `SpaceBlockData` structure directly.
This makes the stored `EvaluatedBlock` data available to users of both,
which will be useful to allow removing `Block`'s accessor methods in
favor of exclusively using EvaluatedBlock.

I'm not entirely sure making SpaceBlockData a public type is wise, but
we haven't really got a stable API yet anyway, and if necessary it can
become a wrapper.