Put raytracing data into a struct, `SpaceRaytracer`.
Using the `ouroboros` library allows us to create a self-referential
struct which can store the `TracingCube`s without requiring another
temporary variable for the `TracingBlock`s they refer to. This will
allow more flexibility in how to schedule raytracing, such as
doing it within an iterator that is returned from a function.