Define the "Cornell box" test scene.
https://en.wikipedia.org/wiki/Cornell_box

As reference for the object sizes and positions, I used the POV-Ray demo
https://github.com/POV-Ray/povray/blob/74b3ebe07c19d556d940ad25518d76426d59ad38/distribution/scenes/radiosity/cornell.pov
but without attempting particular precision or blocks rotated off the grid.

The reason for adding this at all is that I'm looking for test data to
exercise the algorithms (in this case, lighting) and produce
recognizable depictions of the results, without having to spend a lot of
creativity on inventing programmer-art-grade "game content". Yet. Also,
in particular I want to see how my lighting algorithm is doing at
realism -- I know that it will never be especially accurate given its
simplifications, but it's still interesting to see what it does on
well-known test scenes.