Make `VoxelBrush` used by reference, for flexibility.
`embedded_graphics` requires that the pixel type is `Copy`. What I
missed on the first round of design was that, if I'm going to use
references at all (imposing the presence of a lifetime), then I might
as well use `&VoxelBrush` as the pixel type, which is automatically
`Copy`, so the brush itself need not be constrained to use fixed-length
arrays.

Additionally, `VoxelBrush` now stores the blocks in `std::borrow::Cow`,
so it can (if the caller desires) own the blocks; thus a brush can be
`'static` and stored for repeated use easily. (This required adding
implementations of `From` for `Cow<Block>`, which I expect to see
further use.)

We could additionally imagine a brush that _is_ fixed-length like the
old version, and stores color values instead of blocks, so that the
resulting `embedded_graphics::style::Styled` is `'static`. But no uses
for that have come up yet.