Detailed cost approximation for light updates.
Instead of simply counting the number of updates per step, we count
how many cubes were traced through and whether any updates happened.
This way the CPU load is more even.

(We could try to use actual CPU time measurement, instead, but that
has the disadvantage of making the algorithm non-deterministic.)

In testing, there is still lots of stutter, but it is more evenly
distributed and I think it is because updating chunks is too expensive,
not because of the lighting calculations.