Add `BlockDef` and `Block::Indirect`.
These allow storing the entire definition of a `Block` in a `Universe`,
thus making it centrally editable.

We haven't got block change notifications yet, so actually using this
capability won't go very well; the main reason I'm doing this now is in
order to allow the world-generation code to be done in a style that
creates and uses these definitions, rather than passing blocks around
in ad-hoc data structures.