Add details to error types and implement `Error` via `thiserror`.
There's no great reason to do this now, except that it was bugging me
that my errors didn't implement `Error`.

In order for `RefError` to give the name of the object, I had to change
`URef` to have a `name` field which it can access without attempting to
borrow the `RefCell`. It would be better if we could just use the name
field that's in `UEntry`, by way of `UEntry` being outside of the
`RefCell`, but that would require more fiddling with the `owning_ref`
machinery than I want to dig into right now, if it is even possible.