Reuse texture allocator across block changes.
In order to make this convenient, removed the intermediate struct
BlockGLRenderData; SpaceRenderer now holds the texture allocator
directly. This is leading to a concerning accretion of responsibility
to SpaceRenderer, but I think that this will be handled best by making
SpaceRenderer non-luminance-specific (i.e. just something that knows how
to take triangulator algorithms and turn them into an incrementally
updatable render) so it can be tested as a headless algorithm.

An upside of removing BlockGLRenderData is that the block_texture module
is now strictly concerned with providing the texture allocator
implementation.