Separate landscape block construction from usage via BlockDef.
This refactoring makes things messier for now, but it is needed in order
to non-messily support lots of worldgen code using the same blocks, and
in the future for those blocks to have editable/upgradable definitions.

The next step will be to refine the technique for having well-known
blocks, possibly via enums.