Remove `Air` from LandscapeBlocks.
This eliminates a complication: passing `Block::AIR` through the
`BlockProvider` mechanism resulted in an indirect version of it, which
made it impossible to use `PlaceBlock` since it would consider the
cubes already occupied — `Block::Indirect(BlockDef(AIR)) != AIR`.

In the long run, I would like it to be possible to have different
definitions of empty space (underwater, fog, etc.), but for now this
solves the problem and is less code.