Add `GridMatrix` as the discrete analogue of `Matrix4`.
This way, the triangulator does not need to convert to float coordinates
just to do its face rotation. It should also come in handy for
implementing rotated blocks and worldgen.

The reason we aren't simply using `cgmath::Matrix4` is that Matrix4 only
permits float types for its scalar values. Furthermore, this is arguably
correct as a full 4×4 matrix either only applies to 4-element vectors
(with homogeneous coordinates) or implies dividing by the w component
after transforming a 3-element point. By instead implementing a 4×3
matrix which can only express affine transforms, we avoid any questions
of how to divide integers, at the cost of less expressiveness — that we
don't expect to use.