SpaceRendererTodo tracks only existant chunks.
This will prevent storing dirty flags for out-of-range chunks, and also
makes way for tracking multiple dirty flags separately (such as for
storing lighting in a separate 3D texture).

This means losing the `chunk_queue_count` diagnostic because we no
longer have a data structure whose size corresponds to dirty chunks,
and tracking it would be extra work.