Change draw_to_blocks to use a single space.
The old implementation was from before `Block::Recur` had offsets, so
each block had to use a different space.

Note that the tests had to be changed because the space coordinates
changed; the contents should be the same. Also, there may be off-by-one
errors in the bounds calculation; my understanding is that
embedded-graphics currently has some inconsistencies in how it measures
bounding boxes which will be fixed in version 0.7, so I am holding off
on testing precisely until then.