Flatten `all_is_cubes::content::*` except for `palette`.
This makes sense here because these modules have few symbols and
lots of code. As a general principle, I want files to be small enough
to navigate easily, but don't want the public interface to have to
change as more code is written; I did this for the `math` module but
had not previously considered `content`'s structure.