Fix and redefine `GridRotation::to_positive_octant_matrix`.
This addresses two problems. First, the half-open coordinate range
definition was not as useful as it could be, because in _general_
the offset needs to be applied at the time of rotation rather than
the time of matrix construction. Now that we have
`GridMatrix::transform_cube` which understands how the point
designates a unit cube volume, we can do that.

Secondly, the translation components were not being correctly calculated
(using the input axes rather than the output axes) and they now are.
And there's a test.