Use raytracer to implement a 'render to image and exit' function.
This reveals that the raytracer seems to have a brightness problem
(when compared to the luminance/OpenGL renderer) but we can look at
that separately.

Also, I unfortunately didn't find any crate that currently supported
animated PNG output, which is what I had in mind for 'video' output,
but at least the `png` crate has plans for it. For now, I haven't even
added a command line option for the animation output, but if enabled
it does produce a series of separate frame files which could be
assembled into video using `ffmpeg` or similar.