Use greedy meshing algorithm for block triangulations.
The greedy meshing algorithm is described in the article
<https://0fps.net/2012/06/30/meshing-in-a-minecraft-game/>
and in this context, applying it to individual blocks, it has
the effect of removing 100% of the fully-transparent or obscured
texels that we were drawing, and breaking up some textured quads
into separately colored quads.

By not using texture transparency to define surface edges, we
eliminate the seam artifacts on voxel corners. We don't even need
anti-bleed borders inside the textures, because the per-quad texture
coordinate clamping in the fragment shader eliminates them for free.

Future work: Refactor the algorithm into a struct with methods so that
it can be less repetitive and not one excessively long function.