Remove block resolution constraint in triangulator.
Now that we're using a single texture and also breaking quads into
solid-colored regions sometimes, this actually works much better
than the result that occurred with the hack: the only caveat is
that _if_ a surface is textured then some of its texels will be
wrong (but it won't have blatantly wrong geometry).

I also now understand why the previous code was creating holey
geometry: the check for obscuring voxels didn't care that we weren't
actually drawing those voxels.