Bring back `triangulate_space`.
I noticed dead doc links and other references to it, fixed the
references, and then decided that `SpaceTriangulation::triangulate`
is really an awkward phrase which feels more fit for Java (in that it
makes sense when everything is expected to be associated with a class,
but less so when we have free functions). So, `triangulate_space` is
back, with the same signature except for returning `SpaceTriangulation`
instead of taking a mutable buffer.