Compute view position in compute_matrices.
Strictly speaking, it's silly to do a matrix inversion when the view
position is already known and could just be passed to ProjectionHelper.
But that would add a way for the inputs to be wrong, and this should
generally be a once-per-frame calculation.

Because of too many things named `view`, also renamed `view()` to
`view_matrix()`.