Eliminate `Exhibit.footprint` in favor of constructing the Space first.
I realized that `Grid::new_c` is unsound, because it doesn't check for
nonzero size and can't practically be made to, because Rust doesn't
currently support const functions that can fail. I could use a macro
literal strategy like I did with `rgba_const!`, but it seems simpler
in this case to just stop making the footprints separately available
(which simplifies some of the exhibit code anyway).

The disadvantage is that if we ever want to give exhibits the ability
to adapt to their chosen plot in some way, we'll have to bring back
the footprint field (or some generalization) as a function.