Replace `ToGfxVertex` intermediate type with just having one custom vertex type.
While this uses a tiny bit more memory for block triangulations, due to
storing an unused lighting field, it is a significant simplification of
code and explanations.

I also renamed `Vertex` to `LumBlockVertex` because we will probably
want to define a different vertex type for any significant amounts of
non-block drawing, since there are now so many fields in block vertices.