Add `GenError` and `InGenError` for propagating errors from worldgen.
Right now, this has little value since all the worldgen code takes no
parameters, so it will always fail or always succeed. But once we have
parameterized worldgen, or editing commands like “put a structure on
this spot”, if there are any bugs or bad parameter values, there can
be errors that we want to politely show to the user instead of panicking
about.

The _design_ of these errors, on the other hand, is rather dubious.
I'll give them some polish as I go.