Make `VoxelDisplayAdapter` parameterized with the color type and rename to `DrawingPlane`.
The upcoming `embedded-graphics` version 0.7 will replace `DrawTarget`'s
pixel color parameter with an associated type, meaning that the trait
can no longer be implemented more than once for the same type. To
accommodate this, we need to move the color into a parameter on the
concrete type.

Additionally, the name `VoxelDisplayAdapter` is long and not very
informative. (It was chosen in the sense that a “display adapter”
is a device for interfacing with display hardware, and the voxel space
is the “hardware” in this context.) The new name, `DrawingPlane`,
reflects its role of selecting which 2D plane out of the 3D space
is being drawn on.