Apply sRGB curve to terminal colors.
I don't actually like the looks of this, but I am moderately confident
it is more correct, and that if there's something wrong it's wrong
inside the raytracer's color computations (or both the test data and
the luminance renderer).

That said, I should also start doing full-pipeline color tests (load
known data, compare rendering to a reference image on screen).