Visualize textured quad edges/clamping.
I wrote this to debug whether clamping was being set up properly,
and it seemed useful to keep around to be able to look at the geometry.
A more refined version would also put borders on untextured quads,
but we don't currently have the vertex attributes that would be required
for that.