Handle double-wide characters in terminal.
The strategy we use is to measure what the terminal thinks their width
is, by reading the cursor position before and after drawing them.
This presumes that the character will act the same wherever on the
terminal it is, but has the advantage of not requiring any separate
knowledge of Unicode characteristics, which might not be the same as
how the terminal interprets the characters.

One thing missing: if such a character is drawn at the right edge, it
will cause following lines to shift down (for one frame). We should
handle that by resetting the cursor position at the beginning of each
line, and also handle wrapped bottom-right-corner characters and other
glitches by fully repainting the bottom info section.