Remove axes from demo city.
As noted, I'm removing it because it's standing in the way of making
the city look good: it obstructs the central view and is full of
garish colors. These could be customized away, but it might make more
sense to just add more relevant detail to the city's own road
generation.