Recurse in `cursor_raycast` for precise selection targeting.
Note that the outer cube is still used for determining which face was
clicked, which makes it hard to, for example, place blocks adjacent to,
yet floating above, thin blocks. However, the obvious alternative is
to use the voxel face, which would be surprising in its own ways, and
the third option I've thought of is to find the AAB of the block's
selectable voxels and intersect with _those_ faces, which is a more
complex algorithm that would resemble Minecraft's behavior regarding
partial blocks.

In order to make this work, I had to change `cursor_raycast`'s signature
to accept `Ray` instead of `Raycaster`.

I also updated some stale comments.