Have lighting algorithm record “validity” of each cube.
It will be useful for the upcoming smooth lighting to distinguish
between cubes that are visible but dark from lack of light, and cubes
that are "out of bounds" for one reason or another such that their light
values will always be zero and should not be interpolated into.

Because `PackedLight` now has more fields than the packed RGB value,
it no longer implements `From` conversions to other types; there are
methods instead.