Simplify `BlockProvider` to use `Block` instead of `Cow<Block>`.
I forget what I was thinking of at the time, but there seems to be
practically no benefit from using `Cow` here, at least as long as
the lifetime is `'static`, and it creates ugly reborrows.

In the future we might do something like widely using `Rc<Block>`
(or a custom variation thereof that allows constants too), which
would actually reduce copies.