Split desktop::record into modules; add recorder implementation enum.
This is preparation for introducing rendering/export strategies other
than raytracing. `Recorder` is now a struct with common elements and
`RecorderInner` the enum that contains specific implementation details.

The module split puts generic code (options structs and `record_main()`)
in separate files from the guts of specific operations (here, PNG
writing) and the code that handles the variation (staying in the
`record` module).