`#![allow(clippy::collapsible_else_if)]`
With Rust 1.51.0, Clippy split `collapsible_else_if` out of
`collapsible_if`, which was already allowed. I looked at the results of
both and I want to allow both. I use uncollapsed `else if`s deliberately
when the outer `if`'s condition is not semantically related to the inner
`if`'s condition.