Split shader programs into opaque and transparent modes.
It turns out that since we stopped *using* texture alpha testing, there
isn't actually anything mandatory to do differently, but I want to
introduce more expensive transparency which will benefit from not being
run on opaque triangles.

And, this is a good opportunity to get rid of the alpha discard test,
which turns out to be a significant speedup in WebGL mode.

In order to use the split shaders, `SpaceRenderer` has to be taught to
choose and initialize shaders itself.