Implement volumetric transparency for unit cubes.
The alpha value of a block is now taken to mean the opacity of a unit
length's thickness of the material, and increased or reduced as
appropriate for the actual thickness traversed by the ray from the
camera.

This has no understanding of the context of complex shapes, so it will
stop looking consistent if used in a recursive block, but the
triangulator also isn't handling them well at all, so it's not worse
for now.

Future work:
* We should also teach the CPU raytracer about volumetric transparency,
  to keep things consistent.
* It's not clear what to do with lighting values or how; perhaps the
  interior light value should be used exclusively. But, let's hold off
   solving this problem until smooth lighting is otherwise ready.