Move `FrameClock` to new file `apps/time.rs`.
The main motivation for this is that it doesn't really belong in
the `universe` module, which can now be solely about implementing
`Universe`. Additionally, I've been thinking that passing around a
`Duration` isn't the best way to express the information that `step`
methods need (if nothing else, 1/60 s isn't an integer number of
microseconds), and if I do something about that then it'll make sense
to have a file for the new type next to `FrameClock`.