Give EvaluatedBlock an explicit resolution field.
This eliminates the need to rederive it from the voxels array, and
various failure cases resulting from that.

Future work: this will allow us to use smaller arrays when the block
space does not cover the full volume, which will save time and memory
for the case of high resolution text in an otherwise empty block — a
case that looks simple but currently can be very expensive, and which
I want to support for the sake of labels and menus. That will require
adjusting several algorithms to understand smaller-than-full-block
`EvaluatedBlock::voxels` values, so it's not immediately accompanying
this commit.