Add a `Tick` wrapper struct around the timestep value.
This will create room for concepts like "paused" in the future.
(When paused we need to stop some operations but not others, like the
InputProcessor should still be reacting to user input, and the VUI
should still update itself.)

It will also enable using a more fitting time representation than
Duration.