Replace space.sky_color: RGB with space.physics: SpacePhysics.
SpacePhysics will contain all global characteristics such as light and
gravity. For now it just contains the sky color, and a LightPhysics enum
which doesn't have any other options yet.

The goal of this change is to be able to disable light computation for
block spaces, which don't need it. In the long run I expect to add the
option to switch to flood-fill-style lighting, and a gravity vector.