Define intersection_point using ray data instead of a separate intersection.
This allows us to get a clean result that's always within the bounds it
should be, whereas the previous strategy of scaling the original ray
direction is subject to floating-point arithmetic error (or rather,
subject to _different_ errors than the ray stepping algorithm).

I believe `Body` collision resolution misbehaves less with this change,
though I have not spent much time testing it.

Note: This change caused raycast_bench to become slower, but that
appears to be entirely due to making the `RaycastStep` structure larger,
and not due to additional computation. I expect it will be noise in
actual use cases, and it did not slow down the raytracer.