Redesign light rays and use dot product weighting.
Instead of having face-specific sets of light rays, we precalculate
the dot product (or cosine-of-angle) between each ray and each possible
face normal, and use the result as a weight if that face is visible.

Also, there is now a limit on the maximum ray travel (which should
really be a parameter in `LightPhysics` eventually), and I've increased
the max update cost per frame (so cubes can be updated quicker).