Add an error type for general graphics API failure.
The goal of this change is mainly to get rid of various `.unwrap()`s in
favor of explicitly making functions fallible. Future work might put up
a slightly user-friendly error report, or allow recovery in case of e.g.
somehow a single chunk goes bad, but for now it's just pushing the
unwraps higher up the call stack.