Add error types for `Transaction` methods.
The error types have no fields for now. It's not entirely clear
whether there is anything valuable to communicate; it might be useful
to identify conflicts, but we don't want to allocate data that won't be
used since conflicts are expected during normal gameplay. But at least
the error types now implement `std::error::Error`.