`#[uniform(unbound)] view_position`
It appears that GLSL implementations — even within WebGL — differ in
how much dead code analysis they do. On my macOS development machine,
`view_position` was always considered active, but on Linux machines
it gets considered inactive for `BlockPrograms::opaque`, because
it is used in the vertex shader to compute a value which is unused
by the fragment shader.

In principle we should skip trying to pass that value from the CPU
side, but it's cheap to compute and this would increase coupling and
pose a risk of breaking other platforms the next time I make a code
change.