Fix terminal raytracer not resetting color at the right edge.
The problem was confusion between two different `Option` semantics:
"color should be reset to default" versus "do not emit a color".
To fix this, `ColorMode::convert` now accepts `Option<Rgba>` instead of
`Rgba`, so that it can determine which of the two occurs. As a side
benefit, this reduces duplication in `image_patch_to_character`.