Add fuzz_physics target.
This “test” doesn't “pass” yet, but the goal is that a `Body`
moving around in an unchanging `Space` should never trigger the
push-out recovery routine.

This is loosely based on the existing no_passing_through_blocks test.

Future work: we should make the Space generation based on fuzzer input
data, so that small test cases can be produced.