New type `TriangulatorOptions` for all triangulator functions.
This should reduce the need to change miscellaneous call sites when the
triangulator's needs change. Also, ChunkedSpaceTriangulation will be
able to straightforwardly decide when to invalidate the stored meshes
due to graphics option changes (this is not yet implemented, because
I don't have a way to make such changes yet).