Make `Universe` thread-safe (`Send + Sync`).
This does not provide the lock-ordering mechanisms that would be
necessary to perform arbitrary multithreaded access. That is,
deadlock is possible if `URef`s are used indiscriminately from
multiple threads.

The main benefit of making this change is that an entire
`Universe` can be transferred to another thread (this might be useful in
the future for `all-is-cubes -g record`'s performance), and things which
merely *contain* a `URef` at some remove, but have no plans to access
it, are now `Send + Sync`.

Note that this implies that implementations of our extensibility traits
— `Listener` and `Behavior` — must now be `Send + Sync` too, since they
both are reachable through a `Universe`.