Fix clear color not being sRGB, and sRGB rounding.
The clear color is not passed through the fragment shader so doesn't
get the sRGB conversion that's programmed in the shader.

Also, I noticed what the rounding problem in the srgb_round_trip test
was: `<float> as u8` rounds down rather than to nearest, and I had put
`.round()` in some places but not all.