Add `UniverseTemplate::Atrium`.
This is based on the famous Sponza Atrium test scene, but designed to
fit into the voxel grid rather than to necessarily be faithful to the
original model. I've also taken some cues from the real building.

My goals in adding this are to see how my lighting algorithm compares,
and to improve worldgen tools by setting them a complex problem.

So far I have just the rough layout of the building shape and arches,
and some very minimal 'texturing'.