Replace Makefile with the cargo-xtask pattern.
The motivation for this is to be able to systematically run tests
with various feature combinations (e.g. with and without `rayon`),
but for now it is mostly just a translation of the Makefile.

By “the cargo-xtask pattern” I mean <https://github.com/matklad/cargo-xtask>:
a binary defined in Rust code within the same workspace, named `xtask`,
which makes itself available through `.cargo/config`.