`use std::fmt` instead of fully qualifying its items.
This is a little shorter per usage without going to the length
of also importing `fmt::Formatter` and `fmt::Result`, and it's in line
with common style in examples of implementing formatting traits.

I'm also toying with the idea of enabling no_std usage again, which this
would simplify the migration of.