Draft of using fire animation in atrium.
This isn't the “correct” element, as they were supposed to be
flowerpots, but I put a little bit of work into the firepots that
might be reusable anyway.

The fire isn't actually enabled, because having lots of it overwhelms
chunk updating capacity; I intend to fix that by adding special
animated-block rendering modes, but for now this is just some greebling
for the atrium.