Pipeline video recording for speed.
The raytracer already used Rayon to thread tracing individual
pixels/rays, but everything else was single-threaded. Now, each of the
processing stages gets its own thread, so image encoding and universe
stepping can happen at the same time with a 1-element buffer between
each stage.

This results in nearly 100%-of-all-cores CPU usage and a ~3× speedup,
on my MacBook Pro -- if I turn off using “Best” PNG compression.
Evidently that should be an option, at least for when the plan is to
reencode to a proper video codec.