Add `Slot` type in between `Inventory` and `Tool`.
This allows stacks of items/tools to exist. It is incomplete:

* `Inventory::use_tool` will refuse to operate on stacks.
* `InventoryTransaction` doesn't understand adding/subtracting.

However, since there is not yet a way for a stack to be created unless
an inventory is specifically created with one, this doesn't break any
existing in-game behavior.

Now that we have `Slot::Empty`, I've removed `Tool::None`. This is not
strictly necessary, but I think it will be beneficial to exclude "ten
copies of nothing" from existing, and to allow `Inventory` to sharply
distinguish between empty slots and filled ones.