Rename `Tool::PlaceBlock` to `InfiniteBlocks`.
This prepares for having non-infinite block inventory.

(My experience from Minecraft tells me that I don't want to simply have
a “creative mode” where “regular” items become infinite, because that
makes it hard to make purposeful edits to inventories. But perhaps I'll
change my mind on that with experience with this approach.)