Add a camera update to keep the cursor from lagging.
The previous problem was that events happen in this order:
1. InputProcessor updates character state from input.
2. Main loop takes camera from renderer for cursor ray.
3. Renderer updates camera from character.
4. Renderer draws frame using that camera.
So, the displayed cursor target position would always lag behind the
camera projection. Now, they're entirely in sync because we have a
separate camera update between steps 1 and 2.

This is still not a complete solution — ideally, the character state
and the cursor would also be updated before every *click* is processed,
so that user mouse input has sub-frame accuracy. And of course, it is
inelegant to require this extra update call in every main loop.