Stop using trailing_box as collision box.
The problem with trailing_box is that its faces are not computed in the
generic fashion `body.collision_box.translate(body.position)`, so
floating-point error may result in it missing things that would count
as a collision next frame.

Furthermore, we need boxes that are nudged definitely *before and after*
the collision plane — after to detect what we're stopping against, and
before to validate that the position we're stopping *in* is actually
empty space (and not e.g. having a box corner poking into it), and
making aab_raycast provide those boxes seems like it will be more
awkward than not.