Add Swatinem/rust-cache CI action.
This caches the results of building dependencies, which should
save a lot of action CPU time. It won't cache the build that
`cargo install wasm-pack` performs, but I'm still tinkering with that
to see if I can make it happen.

This isn't especially necessary for my very limited use of GitHub
Actions, but I'm spending less of a free resource this way, and
getting answers faster. Hopefully.