Implement recursive collision detection.
Note: this commit is a bit of a mess; it's something I've been stuck
on getting working and I think it will be more useful to have it 'live'
and improve it incrementally than to keep holding it back on a branch.
It at least does not do harm to the non-recursive case.