Rework `Sink` and its callers.
`take_equal` was never used, and the `Iterator` interface was more
awkward than helpful, so I've replaced it with
`Sink` is only used in tests, so there's no point in another dependency
just for making it more powerful. Also, `IndexMap` was a good fit for
the original idea of focusing on the `take_equal()` operation, but
irrelevant when checking the entire list.

The new `drain()` method returns all of the messages in the order they
arrived, so it's a good fit for checking exactly what notifications
were sent.