New crate `all-is-cubes-content` splitting the demo content.
This allows compilation to be pipelined (though not with much total wall
time savings) or skipped (if only testing the main library or only
modifying the demo content), and also means that users interested in the
library can skip the demo content entirely (which might become
especially worthwhile if I end up adding large assets).

For the most part, the entire `content` module is now its own crate.
However, I did not bother splitting the `palette` module, and, the
following items are primarily used in tests, and so remain; all of
these are hidden and reexported:
  `axes`
  `make_some_blocks`
  `make_some_voxel_blocks`
  `make_slab`

Necessary changes to visibility and such:
* The `notnan` macro is now visible at the `all-is-cubes` crate root;
  this seems to be necessary for an exported macro.
* Several items are `#[doc(hidden)]` instead of public; each one now has
  an accompanying TODO comment.
* `BlockAttributes`, `Spawn`, and `LightPhysics::Rays` are not marked
  non-exhaustive. This is undesirable and should be fixed, but it turns
  out that you can't construct non-exhaustive structs even with `..`
  syntax outside of the crate.