Use `tui` library to handle terminal layout.
This is retrofitted on top of the existing usage of crossterm and is not
as clean as it could be, and probably has some bugs not yet revealed.
But it has fixed the bug where the layout is broken if a block's name
contains newlines.

Note that `tui` uses `unicode_width` to measure the number of character
cells a string occupies (and how to split it across the internal buffer),
which is less accurate across terminals than our existing strategy of
writing them and checking what happens to the cursor. Hopefully, it is
close enough for info text purposes (which may have more slop). If not,
maybe we could contribute it to `tui`?