Add `VisitRefs`, for traversing the `URef` graph.
This will be necessary for a variety of features:
* On loading from save files, making connections (circular ones, if
  nothing else, but also simplicity in the general case).
* Garbage collection.
* Detecting and optionally rejecting cross-universe references.
* (Maybe) allowing a subgraph to be constructed with strong references
  before it is inserted into a `Universe`, thus making `BlockBuilder`
  more convenient to use and making it straightforward to add new
  objects from within a transaction.

Note that because of `Arc<dyn Behavior>` used inside of `BehaviorSet`,
there cannot be a mutable version of `visit_refs`. I expect to address
this by giving interior mutability to `URef`, which hopefully will not
be greatly more problematic than their existing ability to _lose_ their
referents.