Add optional `winit` dependency and glue functions.
This is not yet used, but I am planning to experiment with using
`wgpu` or `softbuffer` for some rendering, both of which use `winit`
for window creation. All of the new code and deps are behind the
`winit` feature flag, which is not yet turned on (except by building
with `--all-features` during development).