Rewrite demo universe setup to not involve a callback.
I want to convert universe generation to async (so that it is capable of
cooperative multitasking and cancellable) and writing a function that
accepts an async function is, for now, a bit of a strain on Rust's type
system. Therefore, this eliminates `new_universe_with_space_setup` to
avoid that issue.